requirements.txt
websocket-client==1.7.0
flask